[libc++] Fix the declarative generation of FTMs (#108843)
We were incorrectly computing whether a FTM has been implemented.
Instead of checking whether any version of the FTM is implemented for
the current Standard, we need to make sure that the correct version of
the FTM has been implemented.

As a drive-by fix, also correctly close the file that we load JSON from,
which was forgotten.

NOKEYCHECK=True
GitOrigin-RevId: c4a42f61151c585b80122108d961ebfd0e29544c