[libc++] Remove libc++'s own stdint.h and locale.h (#107436)
These headers are not doing anything beyond the system or compiler
provided equivalent headers, so there's no real reason to keep them
around. Reducing the number of C headers we provide in libc++ simplifies
our header layering and reduces the potential for confusion when headers
are layered incorrectly.

NOKEYCHECK=True
GitOrigin-RevId: aa7f377c965ca79cf3022ddafe6cbd419bd52db5