[libc++] Add a .clang-tidy file to libcxx/src
This disables `readability-identifier-naming` for the source files,
since names don't have to by _Uglified in the source files. We currently
don't enforce clang-tidy in the source files, so this is only useful to
avoid a bunch of warnings when using an editor that shows the results of
clang-tidy.

NOKEYCHECK=True
GitOrigin-RevId: 8f50321256ecc1e7d433ca00baff47bd8a8bc81e