[libc++][TZDB] Improves system time zone detection. (#127339)
On some (Linux) systems /etc/localtime is not a symlink to the time
zone, but contains a copy of the binary time zone file. In these case
there usually is a file named /etc/timezone which contains the text for
the current time zone name.

Instead of throwing when /etc/localtime does not exist or is not a
symlink use this fallback.

Fixes: #105634

---------

Co-authored-by: Louis Dionne <ldionne.2@gmail.com>
NOKEYCHECK=True
GitOrigin-RevId: f796747a17941681c8f806cab4e10570b97b5089